docs: add docs for collection page element
